fix: layout das paginas, organização de código
