+++ src/main/java/br/com/mural/criando/futuro/controller/NoticiaController.java
@@ -1,11 +1,8 @@
 package br.com.mural.criando.futuro.controller;
 
 import br.com.mural.criando.futuro.model.Noticia;
-import br.com.mural.criando.futuro.service.ImgurService;
 import br.com.mural.criando.futuro.service.NoticiaService;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -13,12 +10,9 @@ import java.util.List;
 public class NoticiaController {
 
     private final NoticiaService noticiaService;
-    private final ImgurService imgurService;
 
-
-    public NoticiaController(NoticiaService noticiaService, ImgurService imgurService) {
+    public NoticiaController(NoticiaService noticiaService) {
         this.noticiaService = noticiaService;
-        this.imgurService = imgurService;
     }
 
     @GetMapping("/noticias")
@@ -31,14 +25,9 @@ public class NoticiaController {
         noticiaService.saveNoticia(noticia);
     }
 
-    @PostMapping("/upload")
-    public ResponseEntity<String> uploadImage(@RequestParam("file") MultipartFile file) {
-        try {
-            String imageUrl = imgurService.uploadImage(file);
-            return ResponseEntity.ok(imageUrl);
-        } catch (Exception e) {
-            return ResponseEntity.status(500).body("Erro ao fazer upload da imagem.");
-        }
+    @DeleteMapping("/noticia/{id}")
+    public void deletarNoticia(@PathVariable(value = "id") Long id) {
+        noticiaService.deleteNoticia(id);
     }
 
 }
