--- src/main/java/br/com/mural/criando/futuro/model/CardapioParcial.java
@@ -1,14 +1,0 @@
-package br.com.mural.criando.futuro.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class CardapioParcial {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-}
