+++ src/main/java/br/com/mural/criando/futuro/repository/CardapioRepository.java
@@ -1,6 +1,6 @@
 package br.com.mural.criando.futuro.repository;
 
-import br.com.mural.criando.futuro.model.Cardapio;
+import br.com.mural.criando.futuro.model.cardapio.Cardapio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
