+++ src/main/java/br/com/mural/criando/futuro/repository/HorarioRepository.java
@@ -1,6 +1,6 @@
 package br.com.mural.criando.futuro.repository;
 
-import br.com.mural.criando.futuro.model.Horario;
+import br.com.mural.criando.futuro.model.cardapio.Horario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
