+++ src/main/resources/application.properties
@@ -1,26 +1,21 @@
 spring.application.name=mural-criando-o-futuro
-server.port=8090
 
 # PostgreSQL Database Configuration (with Supabase connection pooler)
-spring.datasource.url=jdbc:postgresql://aws-0-us-west-1.pooler.supabase.com:6543/postgres?sslmode=require
+spring.datasource.url=jdbc:postgresql://aws-0-us-west-1.pooler.supabase.com:6543/postgres?sslmode=require&prepareThreshold=0
 spring.datasource.username=postgres.cegmxoauvphewmbdqqku
 spring.datasource.password=!HvgTie7izWbMUT
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 
-# HikariCP Connection Pool Configuration
-spring.datasource.hikari.maximum-pool-size=10
-spring.datasource.hikari.idle-timeout=600000
-spring.datasource.hikari.max-lifetime=1800000
-spring.datasource.hikari.connection-timeout=30000
-spring.datasource.hikari.minimum-idle=5
-spring.datasource.hikari.pool-name=HikariCP
 
 # Hibernate C3P0 Configuration
 spring.jpa.properties.hibernate.c3p0.max_size=20
 spring.jpa.properties.hibernate.c3p0.min_size=5
 spring.jpa.properties.hibernate.c3p0.timeout=300
 spring.jpa.properties.hibernate.c3p0.idle_test_period=300
+
+spring.datasource.hikari.data-source-properties.cachePrepStmts=false
+spring.datasource.hikari.data-source-properties.prepStmtCacheSize=0
 
 # JPA Configuration (Automatic schema updates)
 spring.jpa.hibernate.ddl-auto=update
