+++ src/main/resources/static/css/pages/inicio.css
@@ -1,7 +1,11 @@
 .hero-image {
     width: 100%;
-    height: auto;
+    max-width: 1200px;
+    height: 400px;
+    object-position: center;
+    object-fit: cover;
     border-radius: 8px;
+    margin: 0 auto;
     box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
