+++ src/main/resources/templates/index.html
@@ -6,7 +6,7 @@
     <title>Creche Municipal Criando Futuro</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&family=Open+Sans:wght@400;600&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="/css/principal.css">
     <link rel="stylesheet" href="/css/pages/inicio.css">
 </head>
@@ -14,7 +14,7 @@
     <div th:replace="~{componentes/navbar}"></div>
 
     <div class="container my-4">
-        <img src="https://via.placeholder.com/1200x400" alt="Imagem da Creche" class="hero-image">
+        <img src="/imagens/creche.png" alt="Imagem da Creche" class="hero-image">
     </div>
 
     <div class="container my-5">
