+++ src/main/resources/templates/noticia.html
@@ -6,7 +6,7 @@
     <title>Notícia - Creche Municipal Criando Futuro</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&family=Open+Sans:wght@400;600&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="/css/principal.css">
     <link rel="stylesheet" href="/css/pages/noticia.css">
 </head>
@@ -18,10 +18,22 @@
             <div class="col-md-8">
                 <h1 class="news-title" th:text="${noticia.titulo}"></h1>
                 <p class="text-muted" th:text="'Publicado em ' + ${noticia.dataPublicacaoFormatada}"></p>
-                <div class="news-images">
-                    <div th:each="imagem : ${noticia.imagens}">
-                        <img th:src="@{${imagem}}" class="img-fluid rounded mb-4" alt="Imagem da Notícia">
+                <div id="carouselNewsImages" class="carousel slide" data-bs-ride="carousel">
+                    <div class="carousel-inner">
+                        <div th:each="imagem, iterStat : ${noticia.imagens}"
+                             class="carousel-item"
+                             th:classappend="${iterStat.index == 0} ? 'active'">
+                            <img th:src="@{${imagem}}" class="d-block w-100 rounded" alt="Imagem da Notícia">
+                        </div>
                     </div>
+                    <button class="carousel-control-prev" type="button" data-bs-target="#carouselNewsImages" data-bs-slide="prev">
+                        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+                        <span class="visually-hidden">Anterior</span>
+                    </button>
+                    <button class="carousel-control-next" type="button" data-bs-target="#carouselNewsImages" data-bs-slide="next">
+                        <span class="carousel-control-next-icon" aria-hidden="true"></span>
+                        <span class="visually-hidden">Próximo</span>
+                    </button>
                 </div>
                 <div class="news-content">
                     <p th:text="${noticia.texto}"></p>
